+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add SystemDependent annotation [#426](https://github.com/nbbrd/java-design-util/issues/426)
 
+### Fixed
+
+- Fix search for format method in Representable* processor [#421](https://github.com/nbbrd/java-design-util/issues/421)
+
 ## [1.9.0] - 2026-03-02
 
 ### Added